Problem: The first table cell is often hanging when paste

Solution:
1. We should not use the bottom-most block all the time, because we cannot tell the difference between a bottom-most cell and a bottom-most paragraph
2. We have two marks to decide whether the first and last node should insert the first and last block as paragraph.  We should try to see whether the first and the last block is like a `paragraph block`, which by this plugin is considered as which have only one bottom-most non-void block.  If they are paragraphs, we add them as paragraphs, and turn the `firstNodeAsText` and `lastNodeAsText` marks to false to prevent other nodes are added as text.  If they are not paragraphs, still turn the two marks to false, to prevent other nodes are added as text.
3. add other nodes as blocks in the middle