@zhujinxuan 2018-01-23
About extra space by `slate-hyperscript` https://github.com/ianstormtaylor/slate/pull/1379, at current stage it is why I drop `invalid block` by (text==='').