@zhujinxuan 2018-01-23
```
yarn run v1.3.2
$ mkdir -p ./tmp && babel-node ./node_modules/.bin/_matcha --reporter ./support/benchmark-reporter ./packages/*/benchmark/index.js > ./tmp/benchmark-comparison.json && babel-node ./support/benchmark-compare



  benchmarks
    html-serializer
      deserialize
        12.87 → 11.67 ops/sec
      serialize
        1039.64 → 1003.90 ops/sec
    plain-serializer
      deserialize
        525.13 → 574.07 ops/sec
      serialize
        11397.69 → 10544.16 ops/sec
    rendering
      normal
        115.92 → 140.41 ops/sec
    changes
      delete-backward
        144107.38 → 136582.10 ops/sec
      delete-forward
        28313.48 → 34791.94 ops/sec
      insert-text-by-key-multiple
        107.18 → 117.17 ops/sec
      insert-text-by-key
        1471.73 → 1364.33 ops/sec
      insert-text
        1348.52 → 1224.49 ops/sec
      normalize
        6280.71 → 6833.89 ops/sec
      split-block
        59.32 → 59.19 ops/sec
    models
      from-json
        216.28 → 293.17 ops/sec (36% faster)
      get-blocks-at-range
        1180311.96 → 1322284.51 ops/sec
      get-blocks
        4185469.91 → 4923430.50 ops/sec
      get-characters-at-range
        877184.52 → 1076663.89 ops/sec
      get-characters
        3211476.15 → 3583504.90 ops/sec
      get-inlines-at-range
        941642.96 → 1095508.54 ops/sec
      get-inlines
        2877468.96 → 3681829.29 ops/sec
      get-leaves
        2284561.71 → 3473889.29 ops/sec (52% faster)
      get-marks-at-range
        963910.17 → 1092999.09 ops/sec
      get-marks
        2840945.79 → 3459446.10 ops/sec
      get-path
        1090384.88 → 1161358.50 ops/sec
      get-texts-at-range
        1011941.44 → 977278.69 ops/sec
      get-texts
        3410363.78 → 3273679.02 ops/sec
      has-node-multiple
        150950.40 → 154523.47 ops/sec
      has-node
        1165108.50 → 1245016.58 ops/sec
      to-json
        2699.07 → 3395.96 ops/sec
      update-node
        18374.91 → 19108.52 ops/sec
```
By caching parentPath, speed is improved