@ianstormtaylor 2018-01-25
@zhujinxuan can you include code sample examples like in #1366, those help a lot in terms of understanding what behavior this changes/adds. And can you also remove the unrelated commits from this PR? Thanks.