@ianstormtaylor 2018-01-26
@zhujinxuan Ah, actually forget what I said. I forgot that I can just look at the new test cases to see that it is following the expected ones.

For the ones you added, we shouldn't have `insert-fragment-undo`, I think we should have regular `insert-fragment` tests, so we can be sure that the output looks right. And then if there are specific cases to cover with `undo`, they can live in `tests/history/undo/...`. 

Does that make sense? Right now it's hard to tell how the change in behavior affects documents, since the tests all have `.undo()` which means `input = output`.