@zhujinxuan 2018-01-27
`insert-fragment-undo` removed; I find it unnecessary to test the undo function while snapshot is added;

Change the `snapshot` option to `snapshotSelection` option